requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
distro==1.9.0
fastapi==0.119.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
jiter==0.11.0
openai==2.3.0
pydantic==2.12.0
pydantic-settings==2.7.1
pydantic_core==2.41.1
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
supabase==2.0.3
requests==2.31.0
uvicorn==0.37.0
python-dotenv==1.1.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.3.0
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
tiktoken==0.7.0
PyPDF2==3.0.1
python-docx==1.1.2
